Write ShEx directly instead of using Jade
Using Jade to build an HTML document with a single <pre> that we then
dump back to plain text with htmltotext is rather convoluted, and it
makes it hard to debug errors if they occur. Instead, write the ShEx
from shexbuilder.js directly. Also fixes one bug: type links with an
empty linkTo would previously result in a syntactically incorrect `()`,
now the line is completely omitted instead.