Select outgoing items as well
Get the data not just of the selected items, but also of the items they
point to, otherwise we can’t do a lot of inference on item-valued
statements (which are the most interesting ones, anyways).

Unfortunately, this makes the whole process a lot slower – the ShEx
export takes well over an hour now.