Remove lib from shexbuilder.js
lib was a remnant of the Jade version of the final ShEx export, where
some function needed to be available to Jade. Now that we export ShEx
directly, it’s just confusing, so convert the lib functions to regular
functions instead.

The comment() function is renamed to formatComment() to avoid a conflict
with the comment variable in processResult().