Correctly format data types in ShEx export
RDF2Graph records which types are data types and which ones are shapes,
and this information is retained almost all the way through the ShEx
export as well, with objects like these in temp/compact.json:

    {
      "@id": "http://www.opengis.net/ont/geosparql#wktLiteral",
      "@type": "DataType"
    }

However, during the “frame” step, the `@type` entry is removed from all
but one of those type objects, and the one that retains the @type entry
identifying it as a DataType is not necessarily the first one to occur
in temp/result.json. Therefore, to correctly identify which types are
data types (`predicate type` in the ShEx) and which are shapes
(`predicate @type` in the ShEx), instead of simply hard-coding a list of
a few data types, we need to first walk the entire result data and
collect all types that are recorded to be data types, and then use that
information when we actually produce the ShEx code.