Mark more files as .PRECIOUS
The .nt files are what we typically validate against, so those
definitely stay. The .shex also shouldn’t be deleted just because the
final build result we requested was the .html. And finally, let’s just
make the .html precious as well, even though I don’t assume it would be
deleted anyways.