Don’t write shapes for dropped types
If the type for a property is dropped from the ShEx output, and not
referenced anywhere else, remove the corresponding shape from the output
entirely, since it is likely no longer necessary. This considerably
reduces the output for certain shape expressions; for example,
us-presidents.shex goes from 5.5k to 4.0k lines.