Ignore properties occurring in less than 10% of instances
When collecting properties for a class (or for all classes at once),
ignore all properties that don’t occur on at least 10% of all instances
for a class. This reduces the amount of noise in the process when a
property is only used by a few items, when we can’t really infer much
information from those few triples.

Dropping a property is fine at least for the ShEx export, since we don’t
produce CLOSED shapes and properties not mentioned in the shape are then
accepted unconditionally. I’m not sure about the implications of this
change for other output formats, though.

The margin of 10% is of course completely arbitrary, and perhaps I’ll do
some more investigations to find a good value later. For now I tried 1%,
10% and 50% (divide by 100, 10, 2), and 10% produced the shortest ShEx.