Fix whitespace in all Java source file
The wildly inconsistent indentation and other whitespace issues in the
Java source code files made them unnecessarily painful to edit. Fixed by
setting up a .dir-locals.el file with some variables intended to
somewhat match the existing style, and then running the following three
Emacs commands on all Java files:

    M-x mark-whole-buffer
    M-x indent-region
    M-x whitespace-cleanup

In addition, two internal tabs in Main.java were cleaned up manually.

When resolving conflicts with this commit, it is probably best not to
attempt to merge the changes, but to apply the “other” version (not from
this commit) and then run the above commands again.