Merge less classes during simplification
Merging any classes that have any common superclass at all is much too
aggressive for Wikidata. For now, go to the other end of the spectrum
and only merge two classes if one is already a superclass of the other.
Perhaps we can explore alternative strategies (like looking for common
superclasses, but not following more than N hierarchy levels) later.