Disable simplification step 7.4.4
During simplification step 7.4.4, all references that have been merged
into a parent class are removed from the child classes. Since the ShEx
exporter no longer has special support for subclasses, this is not what
we want: instead, for now keep the references there. (This is probably
not ideal, but it seems to help for now.)