Filter external references in shexbuilder
If a predicate has more than one type link, and one of them is an
external reference, it’s not very useful to export that link: the only
way we can describe an external reference in ShEx is as “.”, which also
covers all of the other type links along with anything else (modulo
multiplicities).

Furthermore, Wikidata is strongly typed, so if there is an external
reference type link and also some other (non-external) type link on a
predicate, the external type link cannot actually be an external URI
(like an external identifier with an RDF formatter URL set) but must
originally have been an item for which RDF2Graph couldn’t find a wdt:P31
statement, most likely because we did not load statements for it.

Therefore, it makes more sense to drop such type links when exporting.
It means that some data will fail to validate, but the alternative is a
shape that is mostly toothless because the “.” part will match anything
(modulo multiplicities), regardless of the other type links we know of.