Drop shapes for classes with few instances
If a class only has a few (direct or indirect) instances, don’t emit any
type links for it, and completely drop the shape if it’s not otherwise
referenced (like we already drop classes in some other circumstances).

The idea behind this is that if we only have a few instances for a
class, it might not make sense to infer too much information from them,
since we can’t tell which type links are actually typical and which are
just random aberrations from a handful of atypical items. Notice that we
do this after the simplification in RDF2Graph itself, which I think
makes sense: we’ve already given these type links a chance to propagate
through the type hierarchy, we’re just trimming them now afterwards.  (I
feel like this is also related to simplification step 7.4.4, removing
type links which were merged into parent classes, that we currently skip
because ShEx no longer has special support for class relations – but I
haven’t thought this through yet.)

The motivation for this is to reduce the size of the ShEx output, in
order to hopefully be able to validate more data sets without running
out of memory or execution time; however, so far this unfortunately
hasn’t resulted in any improvements for the tests I’ve performed.