Fix escaping in N-Triples conversion
Backslashes and a few escape characters must be escaped in the N-Triples
export. The full list of escape sequences was taken from the N-Triples
grammar [1], except that we don’t escape the single quote because that’s
not necessary (though allowed) inside double-quoted literals.

Specifically, this fixes a bug where the TeX string of Q5310 (\LaTeX)
would result in an error from Fuseki (illegal escape sequence value: L).

[1]: https://www.w3.org/TR/n-triples/#n-triples-grammar