Make Fuseki listen on ephemeral port
Instead of using the default Fuseki port (3030), tell Fuseki to listen
on port 0, which on Linux means that the kernel will allocate a free
port (see the description of ip_local_port_range in man:ip(7)).
Afterwards, we can find out which port Fuseki is listening on using
lsof. Just like before, we do this in a loop and know that Fuseki is
ready as soon as the port is bound.