Move ShEx exporter’s temp dir inside project
The ShEx exporter operates in a temporary directory, and previously it
would use a plain temp/ under the current working directory, which
causes problems when you try to export multiple projects concurrently.
Instead, we can simply move the temporary directory inside the project
directory, so it’s immediately clear which project it belongs to. This
also means we don’t have to explicitly clean it up in the Makefile,
because we’ve already deleted the entire prodect directory.

shexbuilder.js accepts a command line argument for the temporary file
it processes, but for backwards compatibility also falls back to the
previous path if no argument is specified.

With this change, it is fully possible to run more than one inference
process in parallel.