Reimplement TreeNode.calculateSubClassOfIntanceOfCount
The method would previously not count direct instances, which is
presumably a bug (since that count is not exposed otherwise), and would
throw a StackOverflowError when encountering a cycle. Both issues are
fixed with this reimplementation: first collect all children, in a
manner that doesn’t cause problems on cyclic graphs, and then add up
their instance counts.