Fix TreeNode.simplifyStep4 for cyclic graphs
This method would previously throw a StackOverflowError on cyclic
graphs, endlessly running along the cycle of child nodes trying to
simplify them. Instead, we simplify each node on its own, and each child
node is itself responsible for collecting the type links of its parent
nodes (via a new getAllParents() method analogous to getAllChildren()).
As in step 3, this requires that we separate the marking of type links
to be removed from the actual removal of those type links, reusing the
same temporary list of type links to be removed.