Re-enable ShEx subclass support
Shape inheritance is an experimental feature, available in the
on-shape-expression branches of shex.js and the ShEx Primer. The syntax
is slightly different from what RDF2Graph used to target (though shex.js
still supports the old `&` syntax), but apart from that this is fairly
straightforward: we can re-enable the simplification step that trims
type links from subclasses, and record superclasses as being used
instead of being unused.