Catch errors when loading super- and subclasses
The getAllSubClassOfClass.txt query potentially runs against a different
SPARQL endpoint than the other queries (if the --remoteGraphFull option
is used), and for certain graphs and certain classes it can also be a
rather expensive query, so the endpoint might not return a proper
result, but instead perhaps a truncated response or a timeout error. The
JsonParseException resulting from this should not block the rest of the
recovery process, which should be able to proceed with incomplete
subclass information, based on the class relationships retrieved from
the other classes in the graph. Instead, catch any errors that occur
when running this query, log them (with some extra information: which
class caused this error?), but then proceed.