Support BNodes in N-Triples conversion
RDF blank nodes are used in Wikidata to represent “unknown value”, so we
need to support them as well. Previously, they would be serialized like
URLs (i. e. <t1346184840> rather than _:t1346184840), which is more or
less problematic (Fuseki doesn’t seem to complain, but it might lead to
problems down the line).